Add "Antimony" Open source Cryptocurrency
About
Antimony is a free open source peer-to-peer electronic cash system that is completely decentralized, without the need for a central server or trusted parties. Users hold the crypto keys to their own money and transact directly with each other, with the help of a P2P network to check for double-spending. Made as a product demo of a blockchain